Rename ompReduction.hpp to ompReductionComplex.hpp
