+++ src/QMCWaveFunctions/BsplineFactory/SplineC2COMPTarget.cpp
@@ -14,7 +14,7 @@
 #include "spline2/MultiBsplineEval.hpp"
 #include "spline2/MultiBsplineEval_OMPoffload.hpp"
 #include "QMCWaveFunctions/BsplineFactory/contraction_helper.hpp"
-#include "Platforms/OMPTarget/ompReduction.hpp"
+#include "Platforms/OMPTarget/ompReductionComplex.hpp"
 #include "OMPTarget/OMPTargetMath.hpp"
 
 namespace qmcplusplus
